Consolidate IsInGap and CalcGapDeltaZ methods in LArGeometryHelper (hope this is a sensible thing to do) and modify CrossGapsAssociationAlgorithm and CrossGapsExtensionAlgorithm accordingly. Also add a IsInGap3D method for future use. Have fixed a bug in the 'CalcGapDeltaZ' method, which returns the amount of a given region that is made up of gaps.
Noticed that the ClusterCharacterisationAlgorithm uses the GapFraction variable as one of the variables used to separate tracks from showers. Have modified this algorithm to take account of detector gaps. Have also added a 'MaxShowerLength' variable to ClusterCharacterisationAlgorithm, which tags a cluster as a clear track above this length (the default is 10m).