Use the ordered list directly for GetClosestPosition.
Previously, this function could use around 20% of the CPU time in CR
reco (on a small sample, but still). This is because calling
`FillCaloHitList` means we copy over every calo hit into a new list,
which in my tests took around 15% of the total CPU time for 10 events.

Similarly, that list is then thrown away straight after, with all the
deallocations taking another 5%.

Since this function doesn't actually need that copy at all, just add
a variant of the function that uses the ordered calo hit list directly.